docs: ADR 055 (edit-op protocol + sidebar) and overnight build log
Also moves the inline-content helpers out of content-body.tsx into
lib/inline-content.ts (react-doctor: non-component export broke Fast
Refresh) and drops manual memoisation the React Compiler already provides.
